version=12.1.49_dev
# bug fix: load "xxxx" filter "biomolecule 1" broken
# bug fix: font labels 16 bold   does not work (i.e. missing face)

# new feature: color atoms property VECTORSCALE
# new feature: isosurface slab n  -- Erik Wyatt idea
# new feature: isosurface NCI     -- Henry Rzepa idea
#   -- NCIPLOT promolecule calculation for reduced density. Use
#
#        isosurface NCI map NCI
#
#      to plot reduced density mapped with ABS(rho)*SIGN(lambda2)
#      where lambda2 is the middle eigenvalue of the Hessian matrix of
#      promolecular electron density
#
#   "Revealing Noncovalent Interactions", 
#   Erin R. Johnson, Shahar Keinan, Paula Mori-Sanchez, Julia Contreras-Garcia, Aron J. Cohen, and Weitao Yang, 
#   J. Am. Chem. Soc., 2010, 132, 6498-6506. email to julia.contreras@duke.edu
# 
#   "NCIPLOT: A Program for Plotting Noncovalent Interaction Regions"
#   Julia Contreras-García, Erin R. Johnson, Shahar Keinan, Robin Chaudret, Jean-Philip Piquemal, David N. Beratan, and Weitao Yang,
#   J. of Chemical Theory and Computation, 2011, 7, 625-632




Former-commit-id: def71a1e7668ac8285be110e2acc811284fd16d2

+++ Jmol/src/org/jmol/quantum/NciCalculation.java
@@ -166,9 +166,9 @@ public class NciCalculation extends QuantumCalculation implements
       double z2 = zeta2[znuc];
       double z3 = zeta3[znuc];
       //System.out.println(i + " " + znuc + " " + z1 + " " + z2 + " " + z3 + " " + coef1[znuc] + " " + coef2[znuc] +" " + coef3[znuc]);
-      float x = xBohr[ix] - qmAtoms[i].x;
-      float y = yBohr[iy] - qmAtoms[i].y;
-      float z = zBohr[iz] - qmAtoms[i].z;
+      double x = xBohr[ix] - qmAtoms[i].x;
+      double y = yBohr[iy] - qmAtoms[i].y;
+      double z = zBohr[iz] - qmAtoms[i].z;
       double r = Math.sqrt(x * x + y * y + z * z);
       double ce1 = coef1[znuc] * Math.exp(-r / z1);
       double ce2 = coef2[znuc] * Math.exp(-r / z2);
